Split a large wrappers module into several files
